WIP unit test for counting votes
