+++ server.js
@@ -4,6 +4,7 @@ const app = express();
 const path = require('path');
 const bodyParser = require('body-parser');
 const md5 = require('md5');
+const countVotes = require('./public/countVotes.js')
 
 
 app.use(bodyParser.json());
@@ -63,7 +64,7 @@ io.on('connection', (socket) => {
 	socket.emit('statusMessage', 'You have connected.');
 
 	socket.on('voteCast', (id, photo, name) => {
-		countVotes(id, photo, name);
+		countVotes(id, photo, name, app);
 		socket.emit('statusMessage', 'Thanks for voting!');
 		io.sockets.emit('voteUpdate', id, photo, app.locals.votes);
 	});
@@ -76,17 +77,17 @@ io.on('connection', (socket) => {
 	});
 });
 
-const countVotes = (id, photo, name) => {
-	const vote_id = md5(name);
-	const filterVote = app.locals.votes.filter(vote => vote.name !== name);
-	app.locals.votes = filterVote;
-	app.locals.votes.push({
-		vote_id : vote_id,
-		button_id: id,
-		photo: photo,
-		name: name
-	});
-	return app.locals.votes;
-};
+// const countVotes = (id, photo, name) => {
+// 	const vote_id = md5(name);
+// 	const filterVote = app.locals.votes.filter(vote => vote.name !== name);
+// 	app.locals.votes = filterVote;
+// 	app.locals.votes.push({
+// 		vote_id : vote_id,
+// 		button_id: id,
+// 		photo: photo,
+// 		name: name
+// 	});
+// 	return app.locals.votes;
+// };
 
 module.exports = server;
